service/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2023.7.22
charset-normalizer==3.2.0
chex==0.1.82
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
distrax==0.1.4
dm-acme @ git+https://github.com/deepmind/acme@fb87fd53ce7250584dc94c481cba949b857ff255
dm-env==1.6
dm-haiku==0.0.9
#dm-launchpad==0.5.0
#dm-reverb==0.7.2
#dm-sonnet==2.0.1
#dm-tree==0.1.8
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.5.4
google-api-core==1.34.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.31.0
google-cloud-appengine-logging==1.3.1
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.11.4
google-cloud-core==1.7.3
google-cloud-logging==3.1.2
google-cloud-resource-manager==1.10.3
google-cloud-spanner==1.10.0
google-cloud-storage==2.2.1
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.57.0
grpcio-gcp==0.2.2
grpcio-status==1.48.2
grpcio-tools==1.48.2
gym==0.26.2
gym-notices==0.0.8
h5py==3.9.0
idna==3.4
jax==0.4.13
jaxlib==0.4.13
jmp==0.0.4
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.0
ml-dtypes==0.2.0
mock==5.1.0
numpy==1.25.2
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.5
overrides==7.4.0
packaging==23.1
Pillow==10.0.1
portpicker==1.6.0
proto-plus==1.22.3
protobuf==3.20.2
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
random2==1.0.1
readerwriterlock==1.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rlax==0.1.6
rsa==4.9
scipy==1.11.2
Shapely==1.8.5.post1
six==1.16.0
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.0.8
tensorflow-io-gcs-filesystem==0.33.0
tensorflow-probability==0.15.0
termcolor==2.3.0
tf-estimator-nightly==2.8.0.dev2021122109
toolz==0.12.0
trfl==1.2.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.3.8
wrapt==1.15.0
